bux fix

+++ app/src/main/java/com/example/askandteach/utils.java
@@ -9,8 +9,8 @@ public class utils {
     public static DateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd hh:mm");
 
 
-    public static String timestampToDatestring(long timestamp){
-        Date date=new Date(timestamp);
+    public static String timestampToDatestring(float timestamp){
+        Date date=new Date((long)timestamp);
         return dateFormat.format(date);
     }
